Note: notes kept by Robert and Wolfgang Huber about the   progress of the project.
The following are some exploratory scripts to better understand the data and the feasibility of chiptype combination.

runvsn.R: runs the vsn normalization / variance stabilizing transformation on the two data sets (Ferrando, Sabina) and saves the raw data as well as the result into runvsn.RData. For high intensities, that transformation leads to the same as the logarithmic transformation resp. the log-ratio, for low intensities it produces appropriately (according to an error model) shrinked estimators. Purpose: a transformation from one chip type to the other may potentially be easier to estimate than on the original or log-transformed data.

parerrmod.R: auxiliary script to visualize the parameters of the error model as estimated by vsn.

varmean.R: looks at the dependence of the estimated variance (from dCHip) on the mean.

u95u68.R: prints some statistics of the u95u68Common.xls table

duplicates.R: looks at the reproducibility / correlation of probe set means that are
supposedly representing the same gene from THE SAME chip and sample. From this we
hope to understand something about probe sets for the same gene on different chip types.