Add arg exitOnFail to reportTestResult, if TRUE, stop() will be called if the test reports a failure.  I'm not sure this is the right place for this functionality, but we need a way to say: "if this test fails, don't bother continuing".
Add testQCData test that verifies access to the QC data.  We want this
so that we don't error out in getting object names if it isn't
available.

Remove unused commented out code.  Minor reformatting of Chenwei's fixes.

Add fix for unix to set R_LIBS to the temp lib for installation.  We should
probably check to see if R_LIBS is non empty, and prepend and restore :-(