Refactor to have AnnBuilder use a global option to find the URLs for downloading data.
The current default setting is to hit the internet for these
resources.  Functions in downloadSourceData.R allow the creation of a
local mirror of these resources along with utility functions for
reading/writing a config file that can be used to set the global
option in the correct format.  

The options are:

- AnnBuilderPublicDataUrls: used only by the mirroring code
- AnnBuilderSourceUrls: used during package building to find the data
  resources.  Set this to a local mirror to improve performance and
  reduce annoyance.

Added Rd entries for the new functions and data sets.