Change affy package building code to utilize the GO data package.  This makes the packages building *much* faster.
Have loadURL be more verbose and tell us where it is downloading from.  This
will be helpful in debugging the data mirror and helping to remind us that we
are or are not using the mirror.

Have the KEGG code report the URLs that come up with 404 or other errors.