Disable fdp layout on Windows.
The Windows build uses a pre-built graphviz in graphviz-win32/lib.  
This is a version 2.2.1 graphviz built painfully by hand and it does
not include fdp support.  Disabling fdp support allows Rgraphviz
to build again on Windows.

Porting to a recent graphviz lib is proving difficult and will likely
be postponed till after the 1.7 bioc release.