A new version of what is currently readAffySnpAnnotCsv.py Name reflects addition of optional MySQL table creation I left the original code to make a table with ALL affy fields - this can be removed but I like the relative elegance :-)
Newer code is much, much uglier, but builds some derived fields (like end position) to conform with the column names currently specified in the wiki
It produces a much smaller sqlite table and runs
faster as a result.

Note, sqlite indexes are going to be interesting. According to the docs at http://www.sqlite.org/lang_createindex.html, "Every time the database is opened, all CREATE INDEX statements are read from the sqlite_master table and used to regenerate SQLite's internal representation of the index layout." I think this means a substantial delay since it takes quite a while to build an index on half a million rows!