Add compatibility for SQLite-based annotation data packages
This patch allows annaffy to be used with environment and SQLite-based
annotation data packages.  The vignette has been converted to use the
hgu95av2db package for testing purposes.

Since the vignette was used to drive the required changes, there may
be some functionality that is not excercised by the vignette that
needs further patches.