Pre-emptively comment calling of non-existing code.
The master script for the build portion calls a
script to build some tables.
But that script does not exist and as far as I 
can tell never did.
so we comment (for now)
#cd $SRC_BASE/chrlength/script; sh getdb.sh
An email will be send to Nianhua to try and 
clarify this.

This rev also fixes another rm that should probably 
be forced.