Fixed bug that causes duplications of all metadata.
This is commented for now, because these lines 
cause duplications of metadata, and as far as I can
see there is no benefit to these lines.  But I did 
NOT erase them, because Nianhua claimed that theses
lines would somehow allow her to see if somehow 
the sources had changed on her while building.

After close inspection I can see no possible way 
for this claim to be true, the code is a duplication
of itself with no apparent reference to an outside 
source.  But I will only comment becuase at the end 
of the day, this is Nianhuas code, and I just don't 
know *everything*...