Commented out parts of GO that make it HUGE.
There were a couple of tables in the GO db 
which made the files ENORMOUS.  In addition, 
the inclusion of this content in this context
could have led to confusion by end users.  
Basically, users should be able to assume that 
the data they are provided is safe to use. 
Therefore, the inclusion of cross species data  
is inherently risky since a user might for 
example reasonably assume that it was human only.
So this data should really live elsewhere
anyhow.