Added src table checking/modification/err logging
This is a big patch to the build process that adds 
some needed bug source tracking features. These 
will allow me to control that sources do not 
pollute the low level DB tables, and to get a log
file when some badness is attempted to be inserted.

populateTables.R is an example of how the new code
is called (for building the Yeast tables as a 1st 
eg.).  prepareData.R is where the code lives that 
does the checking, data adjusting, table inserting
 and error logging.  It is built upon the 
pdInfoBuilder pkg.  Slight adjustments will allow
me to extend this control to ALL the low level 
package inserts, so that when I build, I will get a 
build each time, and get warning in case there was 
cruft in the sources (which would mean that I would
have to contact the people whose data it was).