Added checking for blank IDs to custom Array src maps
This tiny function has been added to the the prepareData.R
code to just look at the src map files for all "custom"
ID mappings and get rid of blank lines.  This is not 
logged at present.  These files see very little turnover,
and it is very unlikely that I am ever going to want to 
include a blank space as an ID.