Fixes bugs with empty enzyme table in TAIR pkgs, and with OMIM appearing in non-human pkgs.
There was a bug that I introduced last time that 
left us with empty tables for enzyme in the TAIR
pkgs.  There was another bug that resulted in all 
the rat and mouse packages having an omim table.
Fixing the OMIM bug is a little complex, so there 
may have to be a patch for this patch.  Hopefully I 
got all of it this time.