Version: 1.8.2 [2007-08-01] None of the fixes changes the API/function of the BioC release version: o Optimized writeCdfHeader() for memory.  For a CDF with 1,200,000+ units   just writing the unit names would consume 1-1.5GB RAM.  Now it writes   unit names in chunks keeping the memory overhead around 100-200MB. o Made convertCdf() more memory efficient. o BUG FIX: The error message in isCelFile() when the file was not   found was broken.
The main reason for commiting these fixes to release branch is that the release branch works with 2.5.x only and the devel branch with 2.6.x only.  Since we cannot wait for R v2.6.0 to come out, we commit these changes for R v2.5.0 users too.  Note, none of the affxparser API have changed.