Added fly c-some lengths to hand curated DB chrLength.sqlite
This just adds data for fly to the chrlength db.  This needs to be
revisted in the future since these lengths change by a small amount
all the time as new builds release.  So this DB should really be
populated by a script with the latest info for each release.  It's 
not really a big deal though. No biologist in their right mind would 
ever count this as "precisely accurate" for big chromosomes.