Gen 2 pkg bug fixes for YEAST and flies.
The bug fix here for flies is really cosmetic, but helps give the
illusion that the package is the same as last time, but it was done
because a similar bug had to be fixed in the YEAST package (where it
was not really cosmetic), and prevented YEAST from properly listing
the number of REJECTORFs.  Also fixed is the much more serious
common2systematic problem in the YEAST package.