Minor tweaks to go_obsolete.
There is a bug with the counts in go_obsolete that can best be fixed
by dropping the truly useless (and unused) term_id from this table.
Since this ID is one that can onyl cause trouble in the future, it was
planned for phase out anyways.  So now it's gone.