Hi Robert,
I have checked in as far as I could get today, but need to go now. If 
you have time, can you have a look. If you are happy with it, and there 
remain no major issues, I'd be fine if you sent it to Ruth, otherwise 
I'll be happy to continue on this tomorrow.

 Best wishes
	Wolfgang