Adds some crufty parsing scripts in for inparanoid.
This just adds the 1st set of parsing scripts to inparanoid.  These
are only just useful for an initial pass.  These will have to be
written differently since we are going to have to download the entire
lot of them (dozens and dozens of files) and then parse them all into
a set of DBs.