Fixes bugs with initial import script.
This just makes the import script work correctly.  Next up will be to
extend this so that we import all the tables we need in there.  There
are 30 species in the latest build, so this could be troublesome since
each of these needs to have additiona maps to get from protein IDs to
gene IDs in order to be useful...