Adds mapping tables back in, includes script with test query
This puts the code back that includes the mapping tables and also adds
a test script so that we can see if everything is working as expected
or not.  The test script will probably be deleted later, but it's here
for now just because its important.  The download script has been
modified along with the getsrc script to try and get these *.fa files
out of the way a little bit better.