Tests now shows way to keeping full tables
After mapping the new IDs, I would like to keep all the old protein
IDs still so that the inparanoid data is not compromised.  but I also
want to map in the gene IDs when this is possible.  The new example in
test.sql is how I will do this.