Finally we have the beggingings of a human DB.
This puts the 1st of several tables into the bindb_human.sql scrip so
that these tables will be added to the org.Hs.inp.db.sqlite file.
This file will contain 5 maps.  One for each of 5 species that will be
mapped to humans.  These maps will present the gene IDs whenever that
is possible, and will present the other protein IDs the rest of the
time so as to allow them to still be complete.  Therefore I will be
able to make an env in R that will allow me to find the orthologs from
one species to the next.