This addition begins the shift to smaller inparanoid packages
This shift means that each inp package will only contain the base
tables from inparanoid, while the other tables that are needed to get
the IDs mapped over are going to be included in the organism base
packages as extra tables.  Next up I will put that other information
into the other packages.