Changes to inparanoid and adjusts copyscript.
This updated the paths on the copy script for the new /mnt/cpb_anno
mount point.  It also changes the code for inparanoid so that the
terminal databases will have maps that conform to the standard,
whereby the new protein IDs will be mapped to the same internal IDs as
everything else.