Major update of downloads and metadata.
This update is designed to add metadata for ensembl and flybase
sources.  It affects the inparanoid intermediate database and the
chipsrc intermediate DBs that also have some of this supplemental data
injected.  To do this involved revamping the download scripts as well,
so that if at some point in the future, the ensembl or flybase data
needs its own section (which seems likely), it can be moved out of the
inparanoid section easily.