Made changes to schema to try and reduce user confusion.
Changed inparanoid "id" to "inp_id" because this id can be many
different kinds of id, but needs a lable in this context.  It is not
an ID assigned by inparanoid, but it could be so many different kinds
of things, (flybase protein ID to jackson lab gene ID) that it needed
a label to enable description.