Minor bug fixes for some packages.
Found some minor bugs that are preventing the use of secondary and
tertiary sources of data for identifying probes.  This affects a few
probes in a small handful of the less popular packages.  What was
happening was that in these rare cases, the extra data gets used
correctly to match up the probe with the gene of interest, but then
the unigene or alternate ID gets left behind in the genes table which
results in it being mixed in next to entrez gene IDs in the final
package maps...  The correct fix is to weed these non-entrez gene IDs
out of the final packages.  And this will be implemented when I recode
this portion for wide distribution.  This whole pipeline needs to be
refactored and this is at the top of my list for things to refactor.
For now at least we won't have non-entrez gene IDs mixed in with our
entrez gene IDs.