Modifications to schemas.
These are changes to the schema documentation so that they more
precisely align with what was in the sqlite files that are feeding
into AnnotationDbi.  Also, code was added to
AnnotationDbi-internals.Rd so that it will not complain about the man
pages being in a sub-template within AnnotationDbi for the new
packages.