Added terms to describe custom CDFs and custom schemas
Added another term for another alternate CDF schema as well as a term
to describe custom schemas for annotation packages based on foreign
database schemas.  For all standard annotation packages, there is
already a way to determine and inspect that schema inside of
AnnotationDbi, but for foreign schemas there needs to be a term to
indicate that things are based on a different underlying design.