Addition of code to package up intermediate DBs.
This checkin adds the code that will be used to make biannual updates
of the intermediate databases.  These databases will be shipped out to
developers as very large R packages, and will allow us to ensure that
everyone has the correct intermediate database installed in the
correct place.  One package is planned per supported organism.  The
sqlForge code has been modified to find these databases as a default
parameter so that everyone should be able to just use biocLite to get
the latest intermediate DBs (if they want to make annotation packages)
and then to just use the wrapper functions to crank out .db packages
for supported organisms.