Updates to SQLForge and a better error msg for mget.
Part of this update just changes the AnnotationDbi mget message to
look more like what is expected based on what its parents msg looked
like.  The other part overhauls the part of the code that does the
initial mapping for SQLForge so that it will not write out
indermediate files. This step was uneccessary and generally just a bad
idea that was causing bugs on the windows platform.  The new code
should be more robust to cross platform issues.