Bug fixes to affy chip scripts and to ipi
Mostly these changes are to keep the affy chip scripts in sync with
the newer SQLForge files. Some changes were required that do not alter
the end product in any way, but that change the contents of the
intermediate databases somewhat so as to make these units accessible
to being packaged and shipped around as files for db0 packages.  Also,
some more recent changes were needed for the ipi directory since they
changed the schema of their source material somewhat.