Added some comments explaining some legacy stuff in GO.
There is some code in the part of the pipe that builds GO that can
inspire terror if you don't know that its old legacy stuff that is not
presently used for anything.  Nevertheless it might be handy to have
that old code someday, so I added comments for the sake of
"fear-management".