Adds support for a new org package for yeast.
The previous org package for yeast was org.Sc.sgd and was based on the
SGD, which is sensible because that is the standard for that
community.  This package is based on NCBI data and is an org.Sc.eg
package, which should hopefully reduce some of the complaints I get
about there not being an eg based yeast package.  Also, we needed a
package to map the standard yeast ORF IDs to both EG and uniprot IDs,
and neither of these is available at SGD.