Fixed a few more bugs in CNSet-class.Rd file.
The CNSet-class.Rd now contains an ad-hoc examples of creating a CNSet
object and use of accessors to access data. In particular, we
illustrate how such an object can be constructed de-novo, how feature
annotation can be added, and how the totalCopyNumber method can be
applied to an annotated CNSet.