Added a hack to make breadcrumbs work a little better-- the BiocViews link now goes to the release BiocViews page instead of /help/bioc-views (there is no such page).
The only problem with this is that you might be looking at the devel
version of a package home page, and then click on the BiocViews
breadcrumb, and it will take you to the release version of the
biocViews tree widget. 

I don't see a way around that for now. I figure, better a somewhat
relevant page than a broken link.