Use 'tar zcf' instead of 'R CMD build --no-vignettes' in STAGE1 for building the "light" source tarballs.
During STAGE1, we need to build a "light" source tarball for each package.
By "light" here we mean that we don't care about the vignette or \Sexpr
directives in the man pages: trying to execute the code contained in the
vignette and/or the \Sexpr directives is too expensive (it can take a long
time) and, most importantly, it's too risky (it can fail).
Using 'R CMD build --no-vignettes' for building a "light" source tarball
isn't good enough: it will still try to install the package if the man
pages contain \Sexpr directives. Unfortunately, at this time (R 2.14),
there doesn't seem to be any option for turning off the evaluation of the
\Sexpr. So, for now, we build the "light" source tarball "by hand" i.e.
we just use 'tar zcf'.