No longer use Ajax on the biocViews widget page. All data is preloaded. Should speed up page rendering and make it more reliable. WEB-2 is now fixed in the way it was originally specified.
A new step was added to rake processing-- .json files are "converted"
to .js, which just means that "var something = " is prepended and
";" is appended (and in the case of tree.json, the js version is
a hash and not an array). This allows these files to be included
via script tags.

Also, multiple copies of BiocViews.html are created by 
rake pre_compile, and then when the site is built, these will
have the appropriate script tags in them.