Updates comments on OrganismDbi.
Previous comments were a bit specific.  OrganismDbi can be used
generally to integrate a large number of different Annotation
resources together providing that they 1) exist , 2) can be related to
each other via common keys and 3) implement select.  There are some
other restrictions like we want to always have a TranscriptDb object
(for example), but generally a LOT can be integrated here.  Also, none
of the current packages provide access to sequence right now since we
could deduce that from BSGenome objects (based on ranges).  But maybe
that is one for the future?