edgeR 3.1.6
Changing the aveLogCPM() function so that it spreads out the low
abundance values more.  It now gives negative as well as positive
values.

The value for AveLogCPM is now computed just once, and not updated
when the dispersion is re-estimated.  This serves now as a standard
measure of abundance for the purpose of trends and plots.  The use
of the old abundance measure is phased out.