Changed vignette to be in line with this API change, as described by Gordon Smyth:
We no longer permit users to pass an offset matrix as an argument when 
running estimateGLMCommonDisp() on an DGEList object.  Hence

   estimateGLMCommonDisp(y, design, offset=offset)

will give an error if y is a DGEList object.  We require the offset matrix 
to already be in y, e.g.,

   y$offset <- offset
   estimateGLMCommonDisp(y, design)

This was always our intended syntax, but now we make it compulsory.