Since I was writing a long email describing these changes, why not paste it in here:
Here is an overview of the changes. 

Previously I made the following changes:

In the BBS python code:
I get today's date and the SVN revision at which the checkout was done (in STAGE1) and append it to a file called
$BBS_WORK_TOPDIR/svnlog.txt. I also copy it to $BBS_WORK_TOPDIR/svninfo which makes it available via
http to the build nodes. Note that unlike the other files in svninfo, this one is not completely regenerated every day, but contains historical information. In particular, for today's run of STAGE2, we need to know the SVN revision number from yesterday, so we can see if a package has changed.

Also, I make sure a file called $BBS_WORK_TOPDIR/installed_today.txt is *deleted* prior to the beginning of STAGE2.

That's the extent of the changes to the python code.

Previously I also made some changes to the R code (STAGE2.R). For now I am just modifying BBS/3.0/bioc/STAGE2.R, but once things are working I can put my changes into BBS/2.14/bioc/STAGE2.R.

So I'll describe what's new now in STAGE2.R, it's a bit of a radical refactor.

We still have the main structure that the following functions in STAGE2.R are called by python code:

updateInstalledPkgs()
preinstallPkgs()
installPkgDeps()
installMultiArchPkgDeps()

Taking these one at a time, updateInstalledPkgs() has not changed at all.

preinstallPkgs() has not changed, but it calls install.packages2() (which also hasn't changed) which calls install.packages.both() which has changed to install CRAN packages from binary and non-cran packages from source.

installPkgDeps() and installMultiArchPkgDeps() now call a different function, install_needed_pkgs(), to do most of their work. They do not use install.packages2() and install.packages.both(). So there's some possible code duplication to come back to.

So basically install_needed_pkgs() takes an argument 'pkg' and  puts together a list of packages, first, the recursive Depends/Imports/LinkingTo dependencies of pkg. Then it adds in the getSuggestedPackagesToInstall() (this function hasn't changed). Then it adds in 'pkg' itself and uniques the resulting list (deps), then does:

for (dep in deps)
{
    if (needs_install(dep))
        smart_install(dep, INSTALL_opts)
}

So needs_install() has the logic to decide if a package needs to be installed. If it's not a BioC software package, it just looks to see if the package is already installed, or if it is, is a newer version available. If it's a BioC package, it looks to see if there have been svn changes since yesterday (if it can't figure this out, it will default to saying yes, we need to reinstall, to be on the safe side). Also, each time a BioC software package is installed, we append to the installed_today.txt file and we check that before installing, to indicate this package has been installed already today, and avoid redundant installs. (Recall that python removes this file before STAGE2 starts, so it does not have out-of-date content). If needs_install() determines that a package does not need to be installed, it will emit a message containing "* DONE (pkgname)" which tells python that the package installed OK.

smart_install() is actually not all that smart since most of the logic is elsewhere. I should probably rename it.
Anyway, it just installs the package. 

There is also a helper function in there called get_pkg_type_for_pkg() which returns either "source" or
getOption("pkgType") depending on various conditions. Note that we only call this function for packages that are NOT BioC software packages (because if a package IS a BioC software package, we already know we need type "source".


So for example, on linux, get_pkg_type_for_pkg() always returns "source".

On other platforms it returns getOption("pkgType") unless there is a newer source version available.

The exception is if a package comes from one of the extra repos (either ours or Ripley's) in which case we want to install the binary if it's available. 

So smart_install() does the installation and then (for BioC software packages) appends the package name to the installed_today.txt file.