Where we stand with the SOP after our 1st pass through.
This document is a work in progress and will never be 'final'.
Eventually 'success' will be defined by the factoring out of its need
to exist at all.  So this document should lead to the elimination of
it's need to exist.  The seeds for this process have been sown
here. But these things will change as plans change.