changing (again) how memory use is reported:
There are many ways we could do this. In the interest of consistency
between back-ends I've chosen to go with how BatchJobs reports 
memory use on a worker. BatchJobs calls gc() on exit instead of right
after job evaluation so numbers will (usually) be a littler higher
than what we see for SnowParam or MulticoreParam.