1.1.1 :
This version is a major update of DAPAR which contains wrappers for all functions. Tus, DAPAR is now compliant with the Proline suite (http://proline.profiproteomics.fr).

 and the aggregation functions were added

The memory managment in the function mvImage has been optimized

A new plot has been added for the missing values filter tools
Added a function to remove contaminants



* To change the current dataset, tue user has only to select te desired one in the menu
* it is possible to modify the gradient of colors for the correlation matrix
A new barplot to show the proportion of contaminants and reverse. The computation of these proportions has been debugged
* adding a function to create new columns in the protein dataset when the peptides one has been aggregated
* feature to build automatically an ID for the MSnset