Remove "nobj" method for Grouping objects added at commit 113624. It was doing:
  length(unique(unlist(x, use.names=FALSE)))

which is based on the incorrect assumption that each object in the original
collection of objects belongs to at least one group. Doing:

  max(unlist(x, use.names=FALSE))

would have been a slightly more efficient alternative but is also wrong.
There is just no way to compute the nobj() of a Grouping object in general.